+++ src/Analyzers/Core/Analyzers/IDEDiagnosticIds.cs
@@ -133,7 +133,7 @@ namespace Microsoft.CodeAnalysis.Diagnostics
         public const string InvalidSuppressMessageAttributeDiagnosticId = "IDE0076";
         public const string LegacyFormatSuppressMessageAttributeDiagnosticId = "IDE0077";
 
-        public const string RemoveUnnecessarySuppressionForIsExpressionDiagnosticId = "IDE0080";
+        public const string RemoveConfusingSuppressionForIsExpressionDiagnosticId = "IDE0080";
 
         // Analyzer error Ids
         public const string AnalyzerChangedId = "IDE1001";
